refactor(embassy): move the dummy `arch` module one level up
